learning

+++ tut-1/basics/CHARdataType.cpp
@@ -4,7 +4,7 @@ int main()
 {
 // 
 char x= 'a';
-cout<<x;
+cout<<(int)x;
 
 return 0;
